scrypt: keep r as a size_t value

+++ src/libsodium/crypto_pwhash/scryptsalsa208sha256/nosse/pwhash_scryptsalsa208sha256_nosse.c
@@ -67,7 +67,7 @@ blkxor_64(escrypt_block_t *dest, const escrypt_block_t *src)
 }
 
 static inline void
-blkcpy(escrypt_block_t *dest, const escrypt_block_t *src, int len)
+blkcpy(escrypt_block_t *dest, const escrypt_block_t *src, size_t len)
 {
 	int i, L;
 #if (ARCH_BITS==32)
@@ -82,7 +82,7 @@ blkcpy(escrypt_block_t *dest, const escrypt_block_t *src, int len)
 }
 
 static inline void
-blkxor(escrypt_block_t *dest, const escrypt_block_t *src, int len)
+blkxor(escrypt_block_t *dest, const escrypt_block_t *src, size_t len)
 {
 	int i, L;
 #if (ARCH_BITS==32)
